# swagger.yaml
openapi: "3.0.0"
info:
  description: "API para gerenciamento de login"
  version: "1.0"
  title: "Auth"
  contact:
    email: "richard_mbs@hotmail.com"
servers:
  - url: http://localhost:5000
    description: Ambiente local

paths:
  /users:
    post:
      tags:
        - "Usuarios"
      summary: "Criar um novo usuario"
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/users-create"
      responses:
        201:
          description: "Usuario foi criado com sucesso"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  id:
                    type: "number"
                    description: "id do usario criado"
                    example: 1
        409:
          description: "Já existe um usuario cadastrado com o email enviado"
        400:
          description: "foi enviado algum parametro incorreto no payload"

  /users/{user_id}/update-password:
    put:
      tags:
        - "Usuarios"
      summary: "Atualizar a senha do usuario"
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/update-password"
      responses:
        204:
          description: "A senha foi alterada com sucesso"
        400:
          description: "Algum parametro incorreto foi enviado, ou a senha antiga esta incorreta"

  /users/{user_id}/groups/assign:
    post:
      tags:
        - "Usuarios"
      summary: "Atribuir um grupo ao usuario"
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/user-assign-groups"
      responses:
        204:
          description: "O Grupo foi atribuido ao usuario com sucesso"
        400:
          description: "Algum parametro incorreto foi enviado no payload"

  /users/{user_id}/groups/unassign:
    post:
      tags:
        - "Usuarios"
      summary: "Remover um grupo do usuario"
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/user-assign-groups"
      responses:
        204:
          description: "O Grupo foi removido com sucesso do usuario"
        400:
          description: "Algum parametro incorreto foi enviado no payload"
  

  /groups:
    get:
      tags:
        - "Grupos de usuarios"
      summary: "Listar os grupos cadastrados"
      responses:
        200:
          description: "Listagem realizada com sucesso"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/groups"
    post:
      tags:
        - "Grupos de usuarios"
      summary: "Criar um novo grupo de usuarios"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
              - "name"
              properties:
                name:
                  type: string
                  example: "grupo de exemplo"
                  description: "Nome do grupo que sera criado"
      responses:
        201:
          description: "Grupo criado com sucesso"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  id:
                    type: "number"
                    description: "id do grupo criado"
                    example: 1
        409:
          description: "Já existe um grupo cadastrado com o nome enviado"
        400:
          description: "foi enviado algum parametro incorreto no payload"
  /groups/{group_id}:
    delete:
      tags:
        - "Grupos de usuarios"
      summary: "Deletar um grupo"
      responses:
        204:
          description: "Grupo excluido com sucesso"
        404:
          description: "Não existe grupo cadastrado com o id enviado"
  /groups/{group_id}/roles/assign:
    post:
      tags:
        - "Grupos de usuarios"
      summary: "Atribuir uma permissão ao grupo"
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/groups-assign-roles"
      responses:
        204:
          description: "A Permissão foi atribuida ao grupo com sucesso"
        400:
          description: "Algum parametro incorreto foi enviado no payload"
  /groups/{group_id}/roles/unassign:
    post:
      tags:
        - "Grupos de usuarios"
      summary: "Remover uma permissão do grupo"
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/groups-assign-roles"
      responses:
        204:
          description: "A Permissão foi remida do grupo com secesso"
        400:
          description: "Algum parametro incorreto foi enviado no payload"


  /roles:
    get:
      tags:
        - "Permissoes"
      summary: "Lista as permissões cadastradas"
      responses:
        200:
          description: "Listagem realizada com sucesso"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/roles"
    post:
      tags:
        - "Permissoes"
      summary: "Criar uma nova permissão"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
              - "name"
              properties:
                name:
                  type: string
                  example: "permissão de exemplo"
                  description: "Nome da permissão que sera criada"
      responses:
        201:
          description: "Permissão criada com sucesso"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  id:
                    type: "number"
                    description: "id da permissão criada"
                    example: 1
        409:
          description: "Já existe uma permissão cadastrada com o nome enviado"
        400:
          description: "foi enviado algum parametro incorreto no payload"
  
  /roles/{role_id}:
    delete:
        tags:
          - "Permissoes"
        summary: "Deletar Permissões"
        responses:
          204:
            description: "Permissão excluida com sucesso"
          404:
            description: "Não existe permissão cadastrada com o id enviado"


                
components:
  schemas:
    users-create:
      type: object
      required:
        - "email"
        - "password"
      properties:
        email:
          type: string
          example: "Jão@gmail.com"
          description: "email do usuario"
        password:
          type: string
          example: "123456789"
          description: "senha"
    update-password:
      type: object
      required:
        - "old_password"
        - "new_password"
      properties:
        old_password:
          type: string
          example: "123456789"
          description: "antiga senha do usuario"
        new_password:
          type: string
          example: "#123!abc"
          description: "nova senha"
    user-assign-groups:
      type: object
      required:
        - "groups_ids"
      properties:
        groups_ids:
          type: array
          items:
            type: number
            example: 1
            description: "ids dos grupos que serão atribuidos ao usuario"

    #Permissões
    roles:
      type: "object"
      properties:
        id:
          type: number
          example: 1
          description: "id da permissão"
        name:
          type: string
          example: 1
          description: "nome da permissão"

    
    #Grupos
    groups:
      type: "object"
      properties:
        id:
          type: number
          example: 1
          description: "id da do grupo"
        name:
          type: string
          example: 1
          description: "nome do grupo"
    groups-assign-roles:
      type: object
      required:
        - "roles_ids"
      properties:
        roles_ids:
          type: array
          items:
            type: number
            example: 1
            description: "ids das permissões que serão atribuidas ao grupo"